requirements.txt
Flask
flask-bootstrap
pytz
python-dateutil
pandas
plotly
aiogram
paho-mqtt
sqlalchemy